fix compile warning
git-svn-id: http://192.168.20.11/sequoiadb/trunk@43448 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/tools/replay/rplDB2LoadOutputter.cpp
@@ -494,7 +494,7 @@ namespace replay
       {
          UINT64 lastSubmitTime = _monitor->getSubmitTime() ;
          UINT64 currentTime = ossGetCurrentMicroseconds() ;
-         if ( currentTime - lastSubmitTime > _submitInterval )
+         if ( currentTime - lastSubmitTime > (UINT64)_submitInterval )
          {
             return TRUE ;
          }
